Hard code zone creation to use vnic
This will need to be made dynamic, so that the IP can be chosen
at run time in order to support multiple zones. Multiple zones
are not currently supported, so there you go.

We use this vnic so that when we start zonegate in the global zone
ipnat will forward everything on port 22 to the local zone.

Outstanding issues:
- when zonegate starts, we no longer have any access to the global
  zone
- almost all zones vagrant commands require access to the global zone