Zone creation happens in zone_info, not in capability
This allows us to create zones outside of the context of vagrant
capabilities. This is important because capabilities use ssh to
test whether the guest is supported. When ZoneGate is running
but the zone is stopped or destroyed, it may be impossible to ssh
into the system. When ssh is accessible, local zones may fail tests
that only succeed in the global zone.

[resolves #19]