Ensure that zone is created AFTER networks are provisioned
NFS with VirtualBox requires that a public or private network is used. 
Because of how the Network middleware is written, the only way to run code
after it is to register middleware *before* it, but call through to later
middleware before executing code.

This change requires a bit switch in the order of our middleware, and will
likely cause problems with attempts to integrate with guest additions. The
guest additions code executes before networks are configured, so before we
can reasonably expect a local zone to exist.

[refs #9]